rename save state to rmp save state
