+++ src/redux/index.ts
@@ -2,7 +2,7 @@ import { combineReducers, configureStore, createListenerMiddleware, TypedStartLi
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
 import accountReducer from './account/account-slice';
 import appReducer from './app/app-slice';
-import saveReducer from './save/save-slice';
+import saveReducer from './rmp-save/rmp-save-slice';
 
 const rootReducer = combineReducers({
     app: appReducer,
